Get tools now tries to obtain an older release if error obtaining latest.
A list of releases is obtained first, and iteration from newest to oldest
release continues if an error occurs obtaining tools for the current release.

As of release 11.5.0 it appears that VMware no longer supplies the release
package which contained the ISO files, but instead has placed the ISO files
inside the core VMware Fusion.app.