Get tools now tries to obtain iso files from core if package attempt fails.
TODO: When getting tools from core, is it possible to obtain the SIG files?

As before if both package and core get tool attempts fail for the latest
release, iteration continues to the previous release.

As of release 11.5.0 it appears that VMware no longer supplies the release
package which contained the ISO files, but instead has placed the ISO files
inside the core VMware Fusion.app.